refactor(workflow-tasks): move VmOrchestrator into library, drop ToolPaths coupling
